Prevent potential connection leak if initial ping fails. (#70)
Prior to this change, it was possible for the Yamux connection to be
established and not told to shutdown. This is possible if the initial handshake
succeeded, but the subsequent ping failed afterwards.

While not likely an actual problem due to the Yamux stream detecting an I/O
error and triggering a shutdown itself, this commit explicitly ensures that a
shutdown is always signalled.